fix(wave2): close 18 high-severity bugs (concurrency, perf, integrity)
# Concurrency
- po_service.receive_items: SELECT FOR UPDATE on the PO row to serialize
  concurrent receives that would otherwise double-receive the same lines.
- batch_repo.increment_quantity: atomic UPDATE for batch qty (closes lost
  update bug under concurrent receipts of the same batch).
- compliance.submit_to_traces: lock the record row + idempotent re-check.
  A user double-click no longer sends two DDS to TRACES NT.
- trace anchor_event: detect orphan tx_sig from a previous failed commit and
  recover via getSignatureStatuses instead of re-sending the memo (closes
  Solana double-spend on retry).
- custody_repo.increment_anchor_attempt: atomic UPDATE (was read+write race).
- solana_client.tx_exists: new helper used by anchor recovery.

# Performance N+1
- po_service.receive_items: bulk pre-fetch PO lines and Products before the
  receive loop. ~400 queries → 2 for typical receipts.

# Frontend perf
- queryClient: staleTime 30s → 60s, refetchOnWindowFocus disabled by default
  (most data isn't volatile enough to warrant refetch on tab change).

# Data integrity
- User.email/username: now unique PER tenant (composite UQ), not globally.
  Closes a multi-tenancy blocker (admin@empresa.com couldn't exist in two
  tenants). Migration 016.
- BusinessPartner.credit_limit: Integer → Numeric(18,2) for COP precision.
- BusinessPartner.discount_percent: Integer → Numeric(5,2) + CHECK 0..100.
  Migration 078.
- WorkflowState.tenant_id: ondelete CASCADE → RESTRICT (consistency).
- AnchorRequest/AnchorRule: added FK to tenants ON DELETE RESTRICT.
  Migration trace 019.

# Multi-tenancy hardening
- variant_repo.VariantAttributeOptionRepository.get_by_id: tenant_id is now
  REQUIRED (no silent default that could leak across tenants).

# Security / DoS
- inventory imports.csv: 10 MB byte cap + 50k row cap (was unlimited
  read-into-memory — DoS vector).
- compliance certificate download: validate resolved file path is inside
  the certificates base dir (defense in depth against pdf_url tampering).

+++ inventory-service/app/db/models/partner.py
@@ -57,8 +57,10 @@ class BusinessPartner(Base):
 
     # ── Customer-specific ─────────────────────────────────────────────────
     shipping_address: Mapped[dict | None] = mapped_column(JSONB, nullable=True)
-    credit_limit:     Mapped[int]        = mapped_column(Integer, nullable=False, server_default="0")
-    discount_percent: Mapped[int]        = mapped_column(Integer, nullable=False, server_default="0")
+    # Numeric so amounts in COP/USD don't lose decimals.
+    credit_limit:     Mapped[Decimal]    = mapped_column(Numeric(18, 2), nullable=False, server_default="0")
+    # Numeric so 12.5% etc. is preservable; CHECK constraint enforces 0..100 range.
+    discount_percent: Mapped[Decimal]    = mapped_column(Numeric(5, 2), nullable=False, server_default="0")
 
     # ── Supplier-specific ─────────────────────────────────────────────────
     lead_time_days:   Mapped[int]        = mapped_column(Integer, nullable=False, server_default="7")
